Moved little shared part of validator to utils.py
Not completely sure if this is the best decision, but it felt like the shared validation fit well into utils.
